REPO-2309 Don't log update and get as events
The two workflow actions update (currently) only changes
the visual representation of it's child folders, so it
does not cause any structural changes. The 'get' action
is even a read-only action. So there is no need to log
the execution of these actions. It would only clutter the
CMS dashboard with unimportant event logs.