HSTTWO-4732 Support a custom prefix for fully qualified links
When invoking

HstLink#toUrlForm(HstRequestContext requestContext, boolean fullyQualified)
with fullyQualified = true, the HST creates fully qualified links
(for the Page Model API always, for request within the channel mgr
however the links will still be relative to the cms host).

Now, the fully qualified URLs sometimes need a custom prefix
because of for example intermediate proxies. A long supported example
of this is binaries which can be loaded from a cdn. For this, the HST
already has support since long via hst:cdnhost which you can configure
on a hst:virtualhost. When a cdnhost is configured, regardless of
fullyQualified = true or false, the links are prefixed with the cdn host.
We need something similar for HstLinks which are not for binaries.

Therefor, the property hst:linkurlprefix (String) has been introduced
which can be configured on hst:virtualhost and hst:mount level. Note that
hst:cdnhost is only configurable on hst:virtualhost, but since the
hst:linkurlprefix is more likely to be used for Page Model API mounts,
the support can be done per mount (but thus also on virtual host level).

There is an important difference whether the hst:linkurlprefix is
configured on virtualhost or on mount level. Assume the following
setup:

+ localhost
   + hst:root
      + resourceapi
          - hst:linkurlprefix = http://www.example.com

With the above, all HST links for 'resourceapi' will be prefixed
with 'http://www.example.com' if theey are marked as 'fully qualified'.

Binary resources however are links relative to the hst:root mount and
will not get the http://www.example.com prefix! If instead configuring

+ localhost
   + hst:root
      - hst:linkurlprefix = http://www.example.com
      + resourceapi

then binary resources will also get the prefix http://www.example.com.
Just like with cdnhost, when a hst:linkurlprefix is configured, the
links will get the prefix regardless of fullyQualified = true or false.

When there is both a hst:cdnhost AND a hst:linkurlprefix configured,
then binary links get prefixed with the cdn host and non binary
links get prefixed with the linkurlprefix (as non-binary links never get
prefixed with the cdn host)