HSTTWO-4731 Make 'context path' behavior consistent in Page Model API
Before this fix, it could be the case that the PMA (page model api)
would not show the 'context path' for the 'live PMA' but it would
include the 'context path' again for the 'channel manager preview PMA'.

This is inconsistent from the SPA pov. The reason why it behaves like
this is that Channel Manager preview pages which are rendered by the
HST *always* require the context path to be present, even if the matched
mount has

hst:showcontextpath = false

That is because the channel manager preview renders the site over the
domain of the CMS. However, for SPAs, this is not the case, and makes it
unhandy and even seemingly inconsistent from the SPA pov.

Hence, the PMA output should both for live and preview honor the
hst:showcontextpath setting: If it is false, don't show context path
for live and neither for the Page Model API preview