CMS-13246 Back port add (save+edit)Draft workflow
Add saveDraft to workflow. SaveDraft sets the hippostd:transferable
property to true. That indicates that authors are allowed to set the
holder to the current user so authors can collaborate on a draft
document without triggering the validation.
Add editDraft to workflow. EditDraft sets the hippostd:holder property
to the current user and sets the hippostd:transferable to false.

(back port from REPO-2315)