CMS-12852 Use orderBefore directly
Code review revealed an issue with the implementation:
The new item (identifier) is added to the children list two
times if there is a match with a sibling item. First before
the sibling and once at the end, because the container node
already has the new item as a child.

The fix is quite simple: Instead of trying to re-use the
updateOrder method just use Node#orderBefore() directly.
This is also more efficient.

(cherry picked from commit edddb413ae9c1f4bcbbe327c6c4526256e284e97)