CMS-13247 Refactor editor
Take out the business logic to determine the model
for the content of the editor and the model used to
render the diff and the mode ( view, edit, compare ).

There was a lot of code duplication. The models and
the mode depended on 5 boolean values and I had to
introduce a 6th one. The test were unreadable and
unmaintainable.

This refactoring improves that. Furthermore now it's
possible to take the "branch" related logic from
the editor itself as well.

(back port from
CMS-12549 Refactor editor)
(cherry picked from commit 86cfe33464ea03017b9f1358abb4309e2bb839b7)