CMS-13007 [Backport 13.4] Introduce Interceptor annotation to use in Dynamic Beans
A new @Interceptor annotation is introduced. The classes which add
this annotation will be scanned and added to a map. While generating
a bean definition, if the cms type of a field of a document type doesn't
match with the known cms types, than the interceptor map will be
searched to find a corresponding interceptor for the specified cms
type. If an interceptor gets found for the specified type, then the
method generation will be delegated to the interceptor.

(cherry picked from commit a5430fa1e43b6c4632fc35bfe4b58b6357e81926)