CMS-13260 Limit actions for transferable document
A document is transferable when it has been saved as draft.

Do not allow publish, copy, move, delete, request (de) publication or
any branch action when the document is transferable.

Also make sure that the document cannot be saved as draft when there is
a pending request, so that a document cannot be published while the
draft is begin edited.t

(back port from CMS-13091)