CMS-13252 Skip validation when save draft
Skip validation when clicking "Keep draft". When the form is null then
the validation is skipped. See
org.hippoecm.addon.workflow.MenuLink#getForm. That was the least ugly
way to do it. For the server side validation is used a boolean flag,
because another workflow call would become very costly.

Move the saveDraft invokation to the editor to be more consistent.

(back port from
CMS-13094 Skip validation when save draft)