CMS-12841 [Backport 14.2] Log errors in case a PMA request inits an http session
Log an error with stacktrace optionally to see which call created an
http session. For PMA requests, we do not want a request to create an
http session.

Since this improvement must be backported to 14.x, it is possible to
suppress the error logging with an hst config property. In 14.x, the
default will be to suppress error logging for created http sessions

(cherry picked from commit 1876983ce35da1f932ee6005122a781a38021153)