CMS-13109 [Backport 14.2] Support properties which are not represented in ChannelInfo
ChannelInfo settings are being stored below the hst channel info node, for example

hst:workspace/hst:channel/hst:channelinfo
as a set of String, Boolean, Date, etc properties.

When the channel settings are read, at this moment, only the properties
are read which are described by a ChannelInfo interface getter

Next to the primary ChannelInfo class for an hst:channel node we also
support channel info mixin classes.

Any stored property on the hst:channelinfo node which is NOT represented
by a ChannelInfo @Parameter is ignored.

When writing / updating the Channel Settings, any property that is
already stored on the hst:channelinfo node but not represented by a
ChannelInfo @Parameter is even removed.

The above is really unhandy: As a developer, I'd like to be able to
bootstrap or configure channelinfo properties (settings) which are not
editable or meant for the channel settings dialog, and I do not want
these properties to be removed on save and I do want these properties
to be available in the channel properties map.

A direct use case is for example to be able to configure the

org.hippoecm.hst.configuration.channel.PreviewURLChannelInfo_url

on channel info node without requiring a channel info interface for it

(cherry picked from commit ee130baafb41011aae27c5c3270ded6b3b2788aa)