CMS-13113 Save the session if node is moved
In addition, instead of checking the sibling by name
it is better to test if the sibling's parent is equal
to the container node. The name is not unique, so the
node might be an item from another container.

(cherry picked from commit 514a70d45983a65cdda975d249fbcbede007aab8)