CMS-13253 Skip validation for 'Keep draft'
Skip validation and confirmation if the document has been saved as draft,
when the document draft variant has been marked as 'transferable'.

Proceed with the workflow call if the document does not validate, but
the MenuAction was added by a StdWorkflow marked as #isFormSubmitted.

When a MenuAction is added by a StdWorfklow marked as #isFormSubmitted,
the validation is skipped, because the AjaxFormSubmitBehavior is not
added.

Reuse the #isFormSubmitted flag to ignore any errors in validations
triggered by other AjaxFormSubmitBehaviors ( e.g. as a result of trying
to close the the tab or refreshing the browser ).

See org/hippoecm/addon/workflow/MenuLink.java:35

(back port from
CMS-13094 Skip validation for 'Keep draft')
(cherry picked from commit 16a909e9627f1a3ba681443d8f3005c1ae228a45)