CMS-13238 Hide translation action if transferable
Hide the translate actions if the document is a transferable document. A
Document is marked as transferable if the draft variant has the
hippostd:transferable property and the value is set to true.

Remove the following condition to hide the translations:
(unpublished,published) and the published variant is currently viewed.
Such a state does not exist ( anymore ).

(back port from CMS-13137 Hide translation action if transferable)