CMS-12856 [Backport 14] Introduce the hst:page mixin node type
[Backport 14.x experience pages]

This mixin is meant to support experience pages to be stored below
document variants. If this mixin is present on a document variant, a
child node called 'hst:page' of type 'hst:component' is allowed.

Since 'hst:component' is a composite, the following can be stored with
this mixin:

+ mydoc
  - jcr:primaryType = hippo:document
  - jcr:mixinTypes = [hst:page]
  + hst:page
      - jcr:primaryType: hst:component
      - hst:referencecomponent: hst:abstractpages/basepage
      + body:
        - jcr:primaryType: hst:component
        - hst:componentclassname: GenericHstComponent
        - hst:template: detail

(cherry picked from commit 10e264415b088d96f9166a8ae7026e97f5c37009)