CMS-12856 [Backport 14] Support rendering Experience Pages
[Backport 14.x experience pages]

An Experience Page is an HST page that is stored below a document variant,
below the node 'hst:page'. These hst page configurations are read
*on the fly* and the hst page configuration instance is unique for the
request. At the end of the request, it should be possible to garbage
collect it, aka, no reference to it should be kept to it.

To load an hst page on the fly we need two things
1) the compository tree of hst:component nodes stored below the document
variant
2) the HST configuration model to be able to resolve inherited components

Number (2) is needed because an experience page should be able to inherit
for example the header and/or footer from the general hst configuration nodes

Because a request hits the site webapp and the HST in memory model is
maintained and build by the platform webapp, we need to service that is
available in the site webapp, and that can build an hst page model for
an experience page. This service should provide a method to load the hst
page model, and the actual implementation of loading it should be delegated
to the platform webapp. Since the platform webapp should do the loading,
but the thread initiating the loading comes from the site webapp, better
to switch classloader during the loading of the hst page as well (although
most likely it would also work without switching classloader since the hst
page loading does not rely on the current thread its classloader).

Putting it all together, the following flow will be supported

1) A request hits the site webapp and gets matched to a sitemap item
2) After (1) we have a resolved sitemap item. Until now, nothing changed
3) The resolved sitemap item has a method to get the hst page for the request:

public HstComponentConfiguration getHstComponentConfiguration()

When the resolved sitemap item has a relative content path pointing to a
node that has the mixin 'hst:page' and has a node 'hst:page', *then* instead
of the HstComponentConfiguration belonging to the sitemap item, we *create*
on the fly a new HstComponentConfiguration for the current request. This
HstComponentConfiguration is then the component configuration for an
experience page
4) For creating a HstComponentConfiguration from the site webapp, we delegate
to the platform webapp via the ExperiencePageService : This is the cross webapp
service that has a method:

HstComponentConfiguration loadExperiencePage(Node hstPage,
                                     ResolvedSiteMapItem resolvedSiteMapItem);

it supports loading a HstComponentConfiguration on the fly. Since via the
resolvedSiteMapItem the backing hst in memory config model can be accessed,
the experience page creation can use this model.

5) After (4) has been done, the normal HST request processing is done with the
experience page HstComponentConfiguration

The above steps is the pretty straigforward wiring of everything. By far the most
complex part is creating the experience page HstComponentConfiguration by gluing
together the hst:page composite node tree and the referenced hst components in this
hst:page, where the reference HST components are components in the in memory HST
model. What we need to be extremely careful about is that the creation of
Experience Pages should be extremely fast, but that it is also highly concurrent.
Since it uses the HST in memory model, we need to make sure that the creation of
experience pages *never* changes any instance variable of the in memory HST model
objects: if it does so, we get concurrency on the in memory HST model objects which
we do not want since then we should take synchronization into account and performance
goes down. For this reason, some methods in the HstComponentsConfigurationService
has been reshuffled a bit, since for example the 'autocreateReferenceNames' impacts
an instance variable of HstComponentsConfigurationService, and thus cannot be used
for experience page creation.
Another very delicate thing is that the Experience Page HstComponentsConfiguration
needs to *clone* any inherited hst components from the in memory hst model. This
happens in the method

HstComponentsConfigurationService#populateComponentReferences which in turn delegates
to HstComponentConfigurationService#populateComponentReferences

However, since 'populate component references' for an experience page looks up
hst components from the in memory HST Model which obviously already have been
populatee *before*, they should not be populated again. Therefor the flag
'referencesPopulated' has been introduced to avoid 're-population' of already
populated references.

Once the HstComponentsConfigurationService#populateComponentReferences returns, the
Experience Page HstComponentService has cloned (deep copy or merge) all HST in memory
HstComponentService it inherits, thus after that, the entire composite tree
of HstComponentConfigurationService instances is tied to the experience page instance
and not to the HST Model