CMS-12543 Preserve scroll position
Preserve the scroll position when the user selects an element that has a
position in the list that is greater than the number of elements that
fit in the view.

In the previous situation, the selection of an item notified the Page
(and there is only one in the case of the cms). The page re-initialized
and the scroll position was lost.

After this commit the page is not notified, because the data grid of the
listdatatable is not versioned.

See org.apache.wicket.Component#isVersioned