CMS-13238 Collect workflows and show it once
The TranslationWorkflow is defined for nodes of type
hippotranslation:translated, so the TranslationWorfklowPlugin is
instanciated three times. The menu was added three times and some how
almost magically the items were combined.

Combine the available languages so that they are only shown once for
each language without resorting to doubtfull hacks.

Only show the translation menu once and then show the combined
information of the three workflows.

Use a combination of the id of the current user session, the handle id
and the name of the class as the key for a reference model, so that a
model is user, handle and session specific.

Use reference models, because a service cannot be registered and aquired
within the same plugin.

Use reference models, because they are bound to the current context and
thread safety, memory leaks and detaching are taken care of by the
plugin framework.

(back port from
CMS-13137 Collect workflows and show it once)
(cherry picked from commit 1308673ffe16c4409ce24349029ff80360dbadfb)