CMS-13187 [Backport 14.2] Always serialize the primary request document in the PMA
Typically, a request has a 'content bean' associated with the request
based on the URL...or call it primary document / folder.

Currently, there is no guarantee at all that the 'root content' item is
actually serialized in the PMA. Also, from the PMA, there is no way to
'know' which document is the 'root content item'.

Certainly for experience pages, it is not so much a specific component
that references / uses the 'request content bean' and puts it into the
model. Therefor, the 'request content bean' can just be missing in
the Page Model API, certainly for experience pages.

Our Page Model API output currently looks something like (v1.0)

{
  "meta" : {
    "version" : "1.0",
    "branch" : "master"
  },
  "links" : {
    "self" : {
      "href" : "http://localhost/site/spa/resourceapi",
      "type" : "external"
    },
    "site" : {
      "href" : "/",
      "type" : "internal"
    }
  },
  "root" : {
    "$ref" : "/page/uid0"
  },
  "page" : {
    "uid0" : {
      "id" : "r26",
I'd suggest to include next to

"root" : {
    "$ref" : "/page/uid0"
  }
also

"document" : {
    "$ref" : "/page/uid1"
  }
and make sure that the 'pagedocument' always gets serialized into the PMA

If the primary content bean for the request does NOT point to a document
(either to no jcr node or a folder), then we do skip the document
attribute completely

(cherry picked from commit 80c14ae535aff4e68135738be5e36cb96e6fc754)