CMS-13155 [Backport 14.x] Introduce a Resource for experience pages
[Backport to integrate/14.x-experience-pages]

This resource will be used to handle REST requests if the
underlying repository node has the mixin that indicates that
we're dealing with an experience page. The CXFJaxrsHstConfigService
will do the delegation based on the node type.

Note that the XPageContainerComponentResource is a clone of the
ContainerComponentResource. They only differ in the PrivilegesAllowed.
Even the injected service is not different, but that will change
because x-pages need a workflow storage mechanism.

(cherry picked from commit 0b9069e671502c37f4f12887c36b460f055b74d7)