CMS-13259 Save empty date
Save the empty date, by setting the model object to
DateConstants#EMPTY_DATE.

In the DatePickerPlugin a DateFieldWidget is added (
DatePickerPlugin.class:63 ) that is initialized with a
MapEmptyDateToNullModel. The DateFieldWidget adds a YuiDateTimeField or
a YuiDateField. Both use YuiDateTimeField to set the model.

In the previous situation the model was not set if the date field was
cleared, because the date was equal to null.

In the new situation the model object is set to the
DateConstants#EMPTY_DATE, so the model which is a
MapEmptyDateToNullModel will set the delegate to DateConstants#EMPTY_DATE,
so getObject will return null, so the JcrPropertyValueModel will be set
to null and eventually the property will set to null, so the empty value
is saved.

(back port from
CMS-13130 Save empty date )