CMS-13159 [Backport 14.x] children of type hippo:skipdraft skip unpublished <--> draft
[Backport to 14.x Experience Page integrate branch]

We have the use case that a feature needs to be able to write directly
to the unpublished variant, bypassing the draft. However, before these
changes, a commitEditableInstance would always push all properties and
child nodes of the draft to the unpublished, by first deleting all
properties and children from the unpublished variant. Obviously, this
would remove changes made directly to the preview

Hence we now support that direct children below a variant of type
hippo:skipdraft are not copied TO draft from unpublished and not from
draft TO unpublished (and thus also not first removed from unpublished
variant). Between unpublished and published the children of type
hippo:skipdraft are just copied as normally was done. Same goes for
versioning and restoring a version.

(cherry picked from commit 1c557483c490f9115523405b62c42a4dc8842d73)