CMS-13156 [Backport 14.x] Support creating container item in XPage
[Backport to 14.x Experience Pages]

This initial commit only supports creating a container item
in XPageContainerComponentResource#createContainerItem. This is
work in progress. Some methods can be extracted, some code
like 'findNewName' is copied from ContainerComponentServiceImpl.

The crux of this commit is that the user session in general
does NOT have write privileges on the unpublished variant, hence
the workflowSession is used to make the changes. At this moment,
the changes are NOT YET persisted, that is also while some
validation in XPageContainerComponentResourceTest fails (on
purpose).

The persisting of the changes should be handle by invoking the
workflow

documentWorkflow.saveUnpublished();

trigger saving the unpublished, result in an audit log, and
make sure that 'lastmodification timestamp' is updated and
'stateSummary' changed : Of course when the unpublished variant
does not have any changes (this can be checked with

hippoSession.pendingChanges(node, "nt:base", prune);

then the saveUnpublished should not result in any changes.

saveUnpublished should not result in a new version in version history.

After saveUnpublished is supported, the
assertions in XPageContainerComponentResourceTest should inspect
JCR Nodes and assert everything is stored correctly

(cherry picked from commit 79607ef01cbc80bfc1dcd98087a29b7fd5de87dd)