CMS-12556 Render unexpected errors in an XML response
In development mode, the CMS displays the default Wicket error page.

In production mode, we use a custom IRequestHandler to render the cause
of the exception in an XML response:

<error type="lenient">The error message</error>

This means we can simply extract the error message in the "Wicket.Event
.Topic.AJAX_CALL_FAILURE" handler and pass it on to the NavApp. If the
type of the error is set to "lenient", the navapp will render the error
message in a non-blocking way (currently as a snack-bar).