CMS-13108 Render collapsible field title for empty optional compounds
If a compound field is optional and has zero compound nodes it will not
call any of the "populate" methods (as there is nothing to populate),
which results in a field that is missing a title.

Instead, we need to render a CollapsibleFieldTitle that is only visible
if there are zero compound nodes.

Clicking on this title should not rotate the "caret" icon, nor should
it add a CollapseBehavior.

To determine if a compound field has any compound nodes, we rely on the
provider's size() method. In compare mode, this means we have to check
both the oldProvider and the newProvider, in preview&edit mode we use
the default provider.

We have to wait until the super.onBeforeRender() call finishes before
all required providers have been instantiated.