CMS-13280 Introduce retainable property
Add an additional property or mixin: retainable. A retainable document
is a document that can accumulate draft changes by repeated "keep draft"
actions.

The main use case is the following situations:

In the content perspective when a tab is being closed and in the visual
editor when the right side panel is being closed: only for a
non-retainable document the user should be presented with the option to
cancel, discard or save the changes.
Give the user feedback that there are draft changes as an effect of 1 or
multiple "keep draft" actions ( feedback messages and icons ).