CMS-13228 [Backport 14.2] Support a '.' in property place holders in host name
When using a property with a dot ${brc.environmentname} as is currently
the variable name an error is thrown:

ModelLoadingException: Node hst:virtualhost is not allowed to be
'${brc.localhost}'. Only ip-addresses are allowed to have a '.'
in the nodename.

This logic should be executed after property expansion

(cherry picked from commit c3da948d64cbc51ac43dabf0ae09b08d70ff0172)