CMS-13301 Do not call #getObject in #toString
The method LoadDetachableModel#getObject logs itself if the DEBUG
log-level is enabled. This requires the #toString method of class
JcrItemModel (which extends LoadDetachableModel) *not* to call
#getObject, as that will trigger an infinite loop.

This fix prevents this by *not* calling #getPath.

Apart from this, the previous implementation of JcrItemModel#toString
has a side-effect that seems out-of-place: if the model was not
detaching and was not attached, it would be explicitly detached and in
turn saved. This feels like a workaround, so I've removed it, but I'm
not sure what the implications are.