CMS-13227 Open draft if document is tranferable
A document is transferable when it is allowed to "transfer" its owner,
another editor can continue editing the document. In that case the
"editDraft" hint is present and true.

The visual editor should open the draft if the document is tranferable
and otherwise the variant return by obtaineditableInstance.

Introduce the SaveDraftDocumentService interface to be able to leave
the existing unit tests of the DocumentServiceImpl as they are. All
business logic in defined in the AbstractSaveDraftDocumentService. The
DocumentServiceImpl delegates to the JcrSaveDraftDocumentService while
to unit test uses the TestSaveDraftDocumentService. That structure makes
the business logic testable.