CMS-13296 [Backport 14.2] Add preview channel manager PMA assertions
We did not yet have integration test assertions for PMA request for the
channel manager preview. For code reuse, to hst-mock the
MockCmsSessionContext object has been extracted from
AbstractFullRequestCycleTest. Note that also a subtle change was made:

When a full request is mimicked, we should not call setPathInfo but call
setServerPath to mimic an Hst request better: Namely, when a real http
request hits a servlet filter, the pathInfo is always empty and the
servletPath is the entire path after the context path. The HST after
matching sets a fake servletPath (the mount path) which is prefixed
with the 'cms preview prefix' (_cmsinternal) if that is on the servlet
path info. After the new servlet path, the rest of the url is set as
pathInfo. This way, the test mock requests better mimic a real request

With the added tests, we now have PMA assertions for Preview Channel
Manager PMA requests, thus we can validate links are created correctly.
The result is that normal PMA URLs (the non-resources) still contain
_cmsinternal if _cmsinternal was present in the request for the PMA,
while for (container) resources (binaries), never _cmsinternal is
returned in the URL in a Page Model API response

(cherry picked from commit 7d81af16b5eb4a4995ec07c7dad8634b33e80c77)