CMS-13302 Improving and fixing the handling of duplicate values when using .meta:add-new-system-values
This now allows and fully supports multiple duplicate values to be bootstrapped, as well as already existing
duplicate values in the repository.
This also better honors the JCR semantics as technically a multi-valued property really is a kind of list or array, not a set.
While intentionally using duplicate values may in practice not be used much, if at all, at least the behavior now is consistent.