CMS-13291 Move validation to validate methods
Part of the functionality of "Keep draft" is to save a draft without
validations. Part of the validations were done in #writeTo.

Move validations from #writeTo to #validate, so a draft document can be
saved without any validations.

Update unit tests accordingly.

The NodeFieldType#validate method has no node parameter, so the
CompoundContext is used to determine the node. The complexity that that
introduces does not seem to outweigh the small increase in complexity of
the NodeFieldType#validate method.