CMS-13299 [Backport 14.x] Assert that draft or published xpage variant cannot be modified
[Backport 14.x Experience Pages]

Make sure that only the unpublished variant can be modified via the
XPageContainerResource. The check on the right variant is done in the
CXFJaxrsHSTConfigService. If detected that it is the draft or published
variant, the '/hst:exception/clientexception' resource is invoked with
a client exception set on the request.

(cherry picked from commit 7d16187d0f3b2d35caedc0f5e9618c4e063e61d7)