CMS-13321 Ignore skipdraft nodes when comparing variants
If a node of a variant is marked hippostd:skipdraft then
it will not be copied from unpublished to draft (e.g.
by calling obtainEditableInstance). So in a deep equals
comparison of a draft with an unpublished variant all
unpublished descendent nodes with the skipdraft mixin
should be ignored.

(cherry picked from commit b6f4bb652603538f19a6b11153fa7a2613c6beac)