CMS-13321 Only apply skipdraft to children
The hippostd:skipdraft mixin is only applied to
the direct children of variants, so do not skip
comparison of descendents even if the mixin is
present.

After a commitEditableInstance the child nodes with
mixin skipdraft of the unpublished will be moved to
the top. Since the draft does not have these nodes
we have to skip all child nodes of unpublished until
we reach the one without the hippostd:skipdraft mixin.

(cherry picked from commit 42d036cad93a693d3218d89bb4fcc28c3965715a)