CMS-13228 [Backport 14.2] In case of missing environment variable, only log INFO
When there is for example a host node name ${brc_stack}, then in case
there is no system property to replace it, it means that host is just
not active. Logging an INFO level message is sufficient then. This
already happened for the host node names directly below a hostgroup, but
host nodes which had a parent host node by accident logged it on ERROR
level. Fixed that now

(cherry picked from commit 252244326ef5470a4abde16c824df35ec1a64391)