CMS-13325 [Backport 14.x] Support Move Container Items for Experience Pages
[Backport 14.x Experience Pages]

When editing an Experience Page in the CM, you can typically drag
component items between components or reshuffling them within the
same container. Within a container there is little room for confusion.

Dragging between containers which belong explicitly to the XPage also
is straight forward what to expect.

However, drag dropping them between containers where one container is
an explicit container of the Experience Page and the other Container is
of an inherited HST Configuration part, then it becomes complex and
confusing

What if the cms user does not have the right role to modify the HST
config part?
As a result of the action, the Experience Page document would change
AS WELL AS the HST configuration : then if the Experience Page document
gets published but the HST config not yet, the container item might be
duplicate for some time or totally missing for some time (depending on
whether moved to or from the XPage)
For the above reasons, we cannot otherwise than prohibit moving
container items between an XPage and HST Configuration.

This commit covers all kind of interactions of moving container items
within a container, within containers of the same hst config or same
XPage, and between hst config and XPages (which is not allowed)

Apart from that, this commit fixes a set of unit test expectations which
were already broken before as a result of 'experience' folder being
added the the unittest content bootstrap yaml in earlier commit

(cherry picked from commit bd60be0935820ea484de313ffceb42d6fdee2a91)