CMS-13028 [Backport 13.4] Get nodes by abs path in case not found by uuid in JcrListener
- Get nodes by absolute path when the node is not found by uuid in
JcrListener. In some cases, the registered uuid in listener is not
valid, for instance when the document is saved, the uuid of the related
node is changed. In that case, the node is fetched by absolute path.
- Do minor code improvements.

(cherry picked from commit 7da67369621972e5e5d5d5b497d665da079a5aa5)