CMS-13327 [Backport 14.x] Implement all XPAGE container ITEM methods needed in the CM
[Backport 14.x Experience Pages]

XPageContainerItemComponentResource Implements the same methods now as
ContainerItemComponentResource, only now specific for XPages instead of
HST config. The main difference is that for XPages no locking is
involved, the changes are made by a workflow session and the workflow
method #saveUnpublished finishes the job.

Note that for some reason MultivaluedMap (interface) did not work to
deserialize to json to hence had to replace it with MultivaluedHashMap
to get the integration tests to work. Not really sure why it worked
from the CM browser before

(cherry picked from commit 38dfcfcda671eb1f125719192a5c0a0fb589d7dc)