CMS-12769 [Forwardport 14.3] Support null on updating multiple props
A derived data function can return null, instead of Value[] for any
property. If the property being written already exists in the document,
and if it is multiple, the an NPE was thrown.
Changed it so now it writes out an empty array

Cherry picked from a5bfa7b083a8f9df2353e2f1b9e4d2ae23f0dfdf