CMS-13210 for protected resources include CORS headers
The SecureCmsResourceServlet only serves resources IF the request is
authorized, aka belongs to a logged in user. The logged in check is
against the http session

For SPAs being rendered in the Channel Mgr, when fetching resources over
the CMS of the host, subsequent requests from a resource are not allowed
if the first request does not contain certain CORS headers, even if the
request is done with including credentials!

So for example when the first request is
//cms.example.org/cms/angular/hippo-cm/main.css
and that main.css tries
to fetch
//cms.example.org/cms/angular/hippo-cm/font.css
then the second request is not send by the browser if the response of
main.css does not include

 Access-Control-Allow-Origin = cms.example.org
 Access-Control-Allow-Credentials = true

This fix covers that the second request now is allowed if it sends
credentials