CMS-13180 Add a label to the viewport object
Since there is no simple way to add a custom translated label at this
point, we decided to use the ID of a viewport instead. To allow for
a bit of control over it, we will use the ID value as it is configured
on the backend (can be in any case) as a fallback value.