CMS-13357 Don't save session on getWorkflowContext
Do not save the session when #getWorkflowContext is invoked.

If a workflow w1 has modified node n1 and workflow w2 wants to modify
the same node quickly after, an javax.jcr.InvalidItemStateException (
external modification ) is thrown.

To synchronize w2 with the latest changes the session should be
refreshed. The workflow context allows access to the internal workflow
session. However DocumentWorkflow#getWorkflowContext triggers a
session#save. When the method #getWorkflowContext does not trigger a
save action then we are able to access the workflow context and refresh
the internal workflow session.