CMS-13358 Refresh session
Refresh session before save and hints actions. The refresh session
before the save action prevents InvalidItemStateExceptions. The refresh
before the hints operation assures that the hints are returned for the
latest state of the document node that has been persisted.

Close the right side panel before switching to the content editor for a
retainable document, because when the user switches back the side panel
is still open, but no backend call is made.

Save the document when the document has to open in "view" mode, because
the backend for the channel manager performs a "edit draft" operation
after each "keep draft" ( See
org.onehippo.cms.channelmanager.content.document.AbstractSaveDraftDocumentService#saveDraft)
, because the side editor should not close on "keep draft".