CMS-13358 Fix enterprise usage of 'Keep draft'
Prevent the following warning in the logs and malfunctioning.
"[INFO] [talledLocalContainer] 07.05.2020 16:37:45 WARN
http-nio-9080-exec-6 [WpmWorkflowDocument.projectNotFound:65] Illegal
state: document '/content/documents/hap/news/2015/08/the-medusa-news' is
associated with a project with id 'null', but the project does not
exists."

The DocumentWorkflow resolves a "null" project to "master", but the
WpmDocumentWorkflow does not. Add "master" explicitly. The "editDraft"
and "saveDraft" hints do not have to take into account the branch. The
keep draft functionality only reads and modifies properties of the draft
document. That variant is not part of any branch.