CMS-13101 Align document info in document editor
Always show info message in the yellow ribbon in the document editor.

The branchNameSupplier can be redefined in subclasses so they can show
their custom branch name.

The property resolver has been added to facilitate testing. The downside
of this approach is that the place holder functionality of the wicket
StringResourceModel cannot be used. The upside is that the test tests
the full english version of the message and limits the need for
manual/end-to-end testing.

Remove the "draft changes" info section from the
AbstractDocumentWorkflowPlugin, because that is now included in the new
message.

Add info to DocumentWorkflowPlugin, because that plugin is always added
to the page, so the branchInfo message will always appear in the yellow
ribbon in the document editor.

The method getBranchName is meant for extenstion, so that plugins can
add their own branchName. This implementation always returns "master".