CMS-13321 Remove redundant draft check
Now that the code is simplified we no longer have to
make sure that a is always draft. The handling of the
skipDraft mixin is now fully symmetric, so the order
doesn't matter. I modified the test so assert that.

(cherry picked from commit 40bf404431888b706decefb2ae67ceba34f101db)