CMS-13208 [Backport 14.x] Support the CMS User JCR session to be available
[Backport 14.x Experience Pages]

During the HST page rendering in the Channel Manager, we need per HST
Component check whether the user is in the right role to be allowed
to do so. For XPage components the required role is 'hippo:author' and
for HST Config components the required role is 'hippo:channel-webmaster'.
To be able to do so, we need to be able to access the JCR Session of
the logged in CMS user.

Therefor make sure to support having the 'delegated JCR Session' which
is a combination of access rights between the HST preview user and the
CMS user, and make sure the cms user session is still available.

(cherry picked from commit ce1949d5972aafc528fec581a4a2fb26a58adf1a)